Refactor: Rename folder & change to "hover:" styling class instead of useState
