+++ packages/frontend/package.json
@@ -34,6 +34,7 @@
     "react-icons": "^5.0.1",
     "react-loading-skeleton": "^3.4.0",
     "react-select": "^5.8.0",
+    "react-swipeable-list": "^1.9.3",
     "yup": "^1.3.3",
     "zod": "^3.22.4",
     "zustand": "^4.5.2"
