+++ packages/frontend/src/components/index/SearchAlbumList.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import React, { useState } from "react";
+import React from "react";
 import { type Song } from "@prisma/client";
 import { MdAlbum } from "react-icons/md";
 
@@ -16,7 +16,6 @@ const SearchAlbumList: React.FC<SearchAlbumListProps> = ({
   songList,
   filteredSongList,
 }) => {
-  const [activeList, setActiveList] = useState(-1);
   return (
     <>
       {showBand && (
@@ -34,11 +33,7 @@ const SearchAlbumList: React.FC<SearchAlbumListProps> = ({
             <Link
               href={`/album/${album?.toLowerCase().replace(/ /g, "-")}`}
               key={i}
-              className={`${
-                activeList === i ? "bg-[#f5f5f6] shadow-md" : ""
-              } flex items-center rounded-lg border-2 p-3 px-5`}
-              onMouseEnter={() => setActiveList(i)}
-              onMouseLeave={() => setActiveList(-1)}
+              className="flex items-center rounded-lg border-2 p-3 px-5 hover:bg-[#f5f5f6] hover:shadow-md"
             >
               <div className="h-full pe-4">
                 <MdAlbum className="h-full w-[36px]" />
