+++ packages/frontend/src/components/index/SearchBar.tsx
