+++ packages/frontend/src/components/index/SearchLoading.tsx
