+++ packages/frontend/src/components/index/SearchSongList.tsx
@@ -20,7 +20,7 @@ const SearchSongList: React.FC<SearchSongListProps> = ({
   getYoutubeVideoId,
 }) => {
   const router = useRouter();
-  const [activeList, setActiveList] = useState(-1);
+
   return (
     <>
       {searchString.trim() !== "" && filteredSongList.length !== 0 && (
@@ -51,11 +51,7 @@ const SearchSongList: React.FC<SearchSongListProps> = ({
           <Link
             key={i}
             href={`/song/${items.name?.toLowerCase().replace(/ /g, "-")}`}
-            className={`${
-              activeList === i ? "bg-[#f5f5f6] shadow-md" : ""
-            } flex flex-col gap-5 border-b pb-3 sm:flex-row sm:rounded-lg sm:border-2 sm:p-3`}
-            onMouseEnter={() => setActiveList(i)}
-            onMouseLeave={() => setActiveList(-1)}
+            className="flex flex-col gap-5 border-b pb-3 hover:bg-[#f5f5f6] hover:shadow-md sm:flex-row sm:rounded-lg sm:border-2 sm:p-3"
           >
             <div className="relative h-[30dvh] w-full overflow-hidden sm:h-[140px] sm:w-[250px] sm:min-w-[250px] sm:rounded md:h-[165px] md:w-[300px] md:min-w-[300px]">
               <Image
